django_skin_disease_detector/requirements.txt
Django==4.2.7
Pillow==10.0.0
tensorflow==2.15.0
numpy==1.24.3
Django==4.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==10.0.0
numpy==1.24.3
tensorflow==2.15.0
keras==2.15.0
django-cors-headers==4.3.1
python-decouple==3.8
